refactor(keyboard): 抽象Home和Main的相关composable
